add personal wage repository jpa class

+++ nts.uk/pr/proto/nts.uk.pr.proto.infra/src/main/java/nts/uk/ctx/pr/proto/infra/entity/paymentdata/personalwage/PprmtPersonWage.java
@@ -16,7 +16,7 @@ public class PprmtPersonWage extends AggregateTableEntity {
     public PprmtPersonWagePK pprmtPersonWagePK;
 	
 	@Column(name = "END_YM")
-	public BigDecimal endYm;
+	public int endYm;
 	
 	@Column(name = "VAL")
 	public BigDecimal val;
