+++ nts.uk/pr/proto/nts.uk.pr.proto.infra/src/main/java/nts/uk/ctx/pr/proto/infra/entity/paymentdata/personalwage/PprmtPersonWagePK.java
@@ -27,6 +27,6 @@ public class PprmtPersonWagePK implements Serializable {
 	public String pWageCd;
 	
 	@Column(name = "STR_YM")
-	public BigDecimal strYm;
+	public int strYm;
 
 }
